djangocrud/requirements.txt
django==3.2
django-cors-headers==3.7.0
psycopg2==2.8.6
djangorestframework-simplejwt==4.4.0
PyJWT==v1.7.1